+++ src/ChannelServer/Util/GmCommands.Handlers.cs
@@ -48,7 +48,7 @@ namespace Melia.Channel.Util
 			Add("go", "<destination>", this.HandleGo);
 			Add("goto", "<team name>", this.HandleGoTo);
 			Add("recall", "<team name>", this.HandleRecall);
-			Add("recallmap", "", this.HandleRecallMap);
+			Add("recallmap", "[map id]", this.HandleRecallMap);
 			Add("recallall", "", this.HandleRecallAll);
 			Add("clearinv", "", this.HandleClearInventory);
 			Add("addjob", "<job id> [circle]", this.HandleAddJob);
@@ -663,33 +663,6 @@ namespace Melia.Channel.Util
 		}
 
 		/// <summary>
-		/// Recalls multiple characters.
-		/// </summary>
-		/// <param name="sender"></param>
-		/// <param name="target"></param>
-		/// <param name="characters"></param>
-		private void RecallCharacters(Character sender, Character target, Character[] characters)
-		{
-			// Recall each player to target's location.
-			var location = target.GetLocation();
-			foreach (var character in characters)
-			{
-				character.Warp(location);
-				character.ServerMessage("You've been warped to {0}'s location.", target.TeamName);
-			}
-
-			if (sender == target)
-			{
-				sender.ServerMessage("You have called {0} characters to your location.", characters.Length);
-			}
-			else
-			{
-				sender.ServerMessage("You have called {0} characters to target's location.", characters.Length);
-				target.ServerMessage("{1} called {0} characters to your location.", characters.Length, sender.TeamName);
-			}
-		}
-
-		/// <summary>
 		/// Warps all players on the map to target's location.
 		/// </summary>
 		/// <param name="conn"></param>
@@ -700,17 +673,44 @@ namespace Melia.Channel.Util
 		/// <returns></returns>
 		private CommandResult HandleRecallMap(ChannelConnection conn, Character sender, Character target, string command, string[] args)
 		{
-			if (args.Length > 1)
+			if (args.Length > 2)
 				return CommandResult.InvalidArgument;
 
+			Character[] characters = null;
+			
+			// Recall same map.
+			if (args.Length == 1)
+				characters = target.Map.GetCharacters(a => a != target);
+
+			// Recall other map
+			if (args.Length == 2)
+			{
+				// TODO: Once we have support for channels and map servers,
+				//   add warp from other servers and restrict recall to
+				//   channel's max player count.
+
+				// Get map id
+				if (!int.TryParse(args[1], out int mapId))
+					return CommandResult.InvalidArgument;
+
+				var map = ChannelServer.Instance.World.GetMap(mapId);
+				if (map == null)
+				{
+					sender.ServerMessage("Invalid map id.");
+					return CommandResult.Okay;
+				}
+
+				characters = map.GetCharacters(a => a != target);
+			}
+
 			// Check for characters
-			var characters = target.Map.GetCharacters(a => a != target);
 			if (!characters.Any())
 			{
 				sender.ServerMessage("No players found.");
 				return CommandResult.Okay;
 			}
 
+			// Recall each player to target's location.
 			RecallCharacters(sender, target, characters);
 
 			return CommandResult.Okay;
@@ -734,7 +734,7 @@ namespace Melia.Channel.Util
 			//   add warp from other servers and restrict recall to
 			//   channel's max player count.
 
-			// Check characters
+			// Check for characters
 			var characters = ChannelServer.Instance.World.GetCharacters(a => a != target);
 			if (!characters.Any())
 			{
@@ -742,6 +742,7 @@ namespace Melia.Channel.Util
 				return CommandResult.Okay;
 			}
 
+			// Recall each player to target's location.
 			RecallCharacters(sender, target, characters);
 
 			return CommandResult.Okay;
